Core: Cleanup `Geometry2D` and `Geometry3D`
• Migrated all logic & private functions to compilation files
• Majority of header includes removed, now forward-declared
• Converted to namespaces
