+++ core/math/quick_hull.cpp
@@ -30,6 +30,7 @@
 
 #include "quick_hull.h"
 
+#include "core/math/aabb.h"
 #include "core/templates/hash_map.h"
 #include "core/templates/hash_set.h"
