+++ drivers/gles3/shader_gles3.h
@@ -33,6 +33,7 @@
 #include "core/math/projection.h"
 #include "core/os/mutex.h"
 #include "core/string/string_builder.h"
+#include "core/templates/a_hash_map.h"
 #include "core/templates/hash_map.h"
 #include "core/templates/local_vector.h"
 #include "core/templates/rid_owner.h"
